+++ ide-gen-proto.sh
@@ -72,6 +72,7 @@ gen_proto_go checkoutservice
 # gen_proto_cpp currencyservice
 # gen_proto_ruby emailservice
 gen_proto_ts frontend
+gen_proto_ts react-native-app
 gen_proto_js paymentservice
 gen_proto_go productcatalogservice
 # gen_proto_php quote