+++ src/react-native-app/ios/reactnativeapp/reactnativeapp-Bridging-Header.h
@@ -1,0 +1,5 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+//
+//  Use this file to import your target's public headers that you would like to expose to Swift.
+//